refactor: add bromelons
